Added code for Object Detection using YOLOv3
This python notebook file was tested and downloaded directly from Google Colab so running it on local system might need some minor changes.
The code also downloads an image file (open-licenced) by default and runs the detection process on the same file. However, only the relative file location to the image needs to be changed to detect objects in another image.